requirements.txt
alembic==1.13.1
bidict==0.23.1
blinker==1.7.0
build==1.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dacite==1.8.1
Flask==2.3.2
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Mako==1.3.5
MarkupSafe==2.1.5
mypy==1.3.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pluggy==1.5.0
polygon-api-client==1.13.7
psycopg2-binary==2.9.9
pyproject_hooks==1.1.0
pytest==8.2.2
pytest-flask==1.3.0
python-engineio==4.9.1
python-socketio==5.11.3
PyYAML==6.0.1
requests==2.30.0
responses==0.23.1
scikit-learn==1.2.2
scipy==1.13.1
setuptools==70.0.0
simple-websocket==1.0.0
SQLAlchemy==2.0.30
threadpoolctl==3.3.0
types-PyYAML==6.0.12.20240311
types-requests==2.30.0.0
types-urllib3==1.26.25.14
typing_extensions==4.12.1
urllib3==1.26.19
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
pytest-flask